Added in debugger again.
Had to unpick the webpack config to understand what was going on.

Upgraded webpack. Fixed a few things.

Also had to copy/dupe the webpack copy scripts that pull some code
from the inner packages to the src/ route for some silly reason to get
the debugger to load up in debug mode when it didn't before.

Added a few extras to package.json for tsc-watch to give me post
build behaviours to emulate webpack for debugging.