Massive fix to get a couple things working.
Removed the old azure code.

Upgraded libraries to get access to apis for ThemeIcons and the AZ-ext libraries.

Adding in elements for help panel.

A lot of defunct tests removed.

Fixed issues with mocking/sinon with the showQuickPick calls. Something weird happened with mocking and when using direct imports it was fixed v's a global module level export/import call.

Next change to upgrade Typescript version.